host poems project on github page
How can we turn demos into poems? Or perhaps; less demos, more poems.
Inspired by the School for Poetic Computation this animation is based on
a video segment viewed at the fall 2015 salon at SFPC. This commit is
representative of the project's first milestone which is the creation of
the basic rearrangement animation.